# notes-service/src/main/resources/swagger.yml
swagger: '2.0'
info:
  title: PoppyNotes Note API
  description: PoppyNotes Note API
  version: "1.0.0"
# array of all schemes that your API supports
schemes:
  - https
# will be prefixed to all paths
basePath: /api/v1
consumes:
  - application/json
produces:
  - application/json
parameters:
  IdNote:
    name: noteId
    in: path
    description: Unique identifier of a Note.
    required: true
    type: string
  NoteCreate:
    name: note-create
    in: body
    schema:
      $ref: '#/definitions/CreateNote'
    required: true  
  NoteUpdate:
    name: note-update
    in: body
    schema:
      $ref: '#/definitions/UpdateNote'
    required: true
  PinNote:
    name: pin-note
    in: body
    schema: 
      $ref: '#/definitions/PinNote'
    required: true
  IdUser:
    name: userId
    in: header
    description: Unique Identifier of the User requesting his notes.
    type: string
    required: true
  Page:
    name: page
    in: query
    description: Page of notes that's beeing returned.
    default: 0
    required: false
    type: integer
  Items:
    name: items
    in: query
    description: Amount of notes per page. 
    default: 10
    required: false
    type: integer
paths:
  /api/v1/notes:
    get:
      summary: Get All Notes from user
      description: Returns all Notes in a shorter form from a user, paginated.
      operationId: getNotes
      parameters:
        - $ref: '#/parameters/Page'
        - $ref: '#/parameters/Items'
        - $ref: '#/parameters/IdUser'
      tags: 
        - Notes
      responses:
        200:
          description: All notes paginated.
          schema:
            type: array
            items:
              $ref: '#/definitions/OverviewNote'
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
    post:
      summary: Create Note
      description: Creates new Note entry.
      operationId: createNote
      parameters:
        - $ref: '#/parameters/NoteCreate'
        - $ref: '#/parameters/IdUser'
      tags: 
        - Notes
      responses:
        201:
          description: Created Note.
          schema:
            $ref: '#/definitions/CompleteNote'
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
    put:
      summary: Update Note
      description: Updated a Note entry.
      operationId: updateNote
      parameters:
        - $ref: '#/parameters/NoteUpdate'
        - $ref: '#/parameters/IdUser'
      tags: 
        - Notes
      responses:
        200:
          description: Note updated.
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
  /api/v1/notes/pin: 
    put:
      summary: Pin Note
      description: Pin a Note.
      operationId: pinNote
      parameters:
        - $ref: '#/parameters/PinNote'
        - $ref: '#/parameters/IdUser'
      tags: 
        - Notes
      responses:
        200:
          description: Note pinned/unpinned.
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
  /api/v1/notes/{noteId}:
    get:
      summary: Get Note
      description: Fetches Note with defined Id.
      operationId: findNote
      parameters:
        - $ref: '#/parameters/IdNote'
        - $ref: '#/parameters/IdUser'
      tags: 
        - Notes
      responses:
        200:
          description: Note entity.
          schema:
            $ref: '#/definitions/CompleteNote'
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
    delete:
      summary: Delete Note
      description: Deletes a Note with defined Id.
      operationId: removeNote
      parameters:
        - $ref: '#/parameters/IdNote'
        - $ref: '#/parameters/IdUser'
      tags: 
        - Notes
      responses:
        200:
          description: Note deleted.
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
definitions:
  PinNote:
    type: object
    properties:
      pinIt:
        type: boolean
        description: Boolean if the note should be set pinned or not
      noteId:
        type: string
        description: Unique Identifier of the Note.
  CreateNote:
    type: object
    properties:
      title:
        type: string
        description: Title of the Note.
      content:
        type: string
        description: Text content of the Note (the actual note).
      userHash:
        type: string
        description: Unique Identifier of the user that created that note.
      lastEdit:
        type: string
        format: date
        description: Date the user last edited the note.
  UpdateNote:
    type: object
    properties:
      id:
        type: string
        description: Unique Identifier of the Note.
      title:
        type: string
        description: Title of the Note.
      content:
        type: string
        description: Text content of the Note (the actual note).
      userHash:
        type: string
        description: Unique Identifier of the user that created that note.
      lastEdit:
        type: string
        format: date
        description: Date the user last edited the note.
      pinned:
        type: boolean
        description: Pin for important messages to be shown always on top. 
  CompleteNote:
    type: object
    properties:
      id:
        type: string
        description: Unique Identifier of the Note.
      title:
        type: string
        description: Title of the Note.
      content:
        type: string
        description: Text content of the Note (the actual note).
      userHash:
        type: string
        description: Unique Identifier of the user that created that note.
      lastEdit:
        type: string
        format: date
        description: Date the user last edited the note.
      pinned:
        type: boolean
        description: Pin for important messages to be shown always on top.
  OverviewNote:
    type: object
    properties:
      id:
        type: string
        description: Unique Identifier of the Note.
      title:
        type: string
        description: Title of the Note.
      content:
        type: string
        description: Text content of the Note (the actual note).
      lastEdit:
        type: string
        format: date
        description: Date the user last edited the note.
      pinned:
        type: boolean
        description: Pin for important messages to be shown always on top.
  Error:
    type: object
    properties:
      message:
        type: string
      fields:
        type: object
        description: Invalid fields.
        additionalProperties:
          type: string