# notes-ui/src/app/service/model/completeNote.ts
/**
 * PoppyNotes Note API
 * PoppyNotes Note API
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface CompleteNote {
    /**
     * Unique Identifier of the Note.
     */
    id?: string;
    /**
     * Title of the Note.
     */
    title?: string;
    /**
     * Text content of the Note (the actual note).
     */
    content?: string;
    /**
     * Unique Identifier of the user that created that note.
     */
    userHash?: string;
    /**
     * Date the user last edited the note.
     */
    lastEdit?: string;
    /**
     * Unique Initialization Vector for note encryption.
     */
    initVector?: string;
    /**
     * Pin for important messages to be shown always on top.
     */
    pinned?: boolean;
}